Rename tests::async_fd::fixed to direct
Matches the name we're using for direct descriptors.
