+++ tests/async_fd.rs
@@ -9,7 +9,7 @@ use util::{is_send, is_sync};
 
 #[path = "async_fd"] // rustfmt can't find the files.
 mod async_fd {
-    mod fixed;
+    mod direct;
     mod fs;
     mod io;
     mod net;
